requirements.txt
requests
lxml
autopep8
python-dotenv
python-decouple
psycopg2
sqlalchemy
numpy
pandas
pyprojroot
pyhere